Add code to detect duplicate guid in harvest job
Without this commit a harvest job with 2 sources using the same guid
would have the effect of the newer source overwriting the old source.

Informing the user that multiple guid are found for the same job should
provide the user with more information as to why a dataset is missing.

In order to change the source for a dataset the publisher will need
to delete the existing dataset and reharvest it.