Revert to the old implementation for Bbox based searches
The approach using Solr's own BBoxField turned out to be not optimal as
it required the use of LocalParams. This is fine if the query is used in
`fq` but CKAN core does not allow it on `q` (unreasonably IMO). Long
story short the old implementation is a bit more verbose but supports:

* Proper spatial sorting
* Combining queries in `q` with a spatial filter
* Antimeridian queries handling

It doesn't support Point geometries, but users can use
`solr-spatial-field` if they need them